+++ changelog.md
@@ -1,2 +1,4 @@
-## Fixed:
-- detection for disconnected user
+## New:
+- function get-known-players() to lists all players who have ever played on a created bedrock-server.
+- function add-player() to complete your player-list yourself.
+- function update-permission() to edit permissions.json for bedrock-server.